Backend/requirements.txt
anthropic==0.49.0
python-dotenv==1.0.0
amadeus==8.0.0
airportsdata==2023.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
flask==2.3.3
flask-cors==4.0.0 